Add /start handler to start job
