+++ main.go
@@ -1,17 +1,14 @@
 package main
 
 import (
-	"bytes"
 	"crypto/tls"
 	"encoding/json"
-	"errors"
 	"flag"
 	"fmt"
 	"github.com/leoleovich/go-gcodefeeder/gcodefeeder"
 	log "github.com/sirupsen/logrus"
 	"io/ioutil"
 	"net/http"
-	"net/url"
 	"os"
 	"time"
 )
@@ -100,6 +97,7 @@ func main() {
 		daemon.config.Listen = "[::1]:8888"
 	}
 	http.HandleFunc("/info", daemon.InfoHandler)
+	http.HandleFunc("/start", daemon.StartHandler)
 	go func() { log.Fatal(http.ListenAndServe(daemon.config.Listen, nil)) }()
 	log.Debug("Started http server on ", daemon.config.Listen)
 
@@ -107,7 +105,7 @@ func main() {
 	daemon.gizmostatusfile = gizmostatusfile
 	daemon.jobfile = jobfile
 
-	ie := &InternEnpoint{
+	daemon.ie = &InternEnpoint{
 		Api_app: daemon.config.InternEnpoint.Api_app,
 		Api_key: daemon.config.InternEnpoint.Api_key,
 		Api_uri: daemon.config.InternEnpoint.Api_uri,
@@ -117,34 +115,32 @@ func main() {
 		job:         &Job{},
 	}
 
-	if err := ie.reschedule(); err != nil {
+	if err := daemon.ie.reschedule(); err != nil {
 		log.Error("reschedule failed: ", err)
 	}
 	for {
 		select {
 		case <-daemon.timer.C:
 			daemon.timer.Reset(pollingInterval)
-			if err = ie.reportStat(); err != nil {
+			if err = daemon.ie.reportStat(); err != nil {
 				log.Error(err)
 			}
 			log.Info("My status is: ", daemon.job.Status)
 
 			switch daemon.job.Status {
 			case "Waiting for job":
-				if err = ie.nextJob(); err != nil {
+				if err = daemon.ie.nextJob(); err != nil {
 					log.Error(err)
 					break
 				}
-				daemon.job.Id = ie.job.Id
-				daemon.job.Filename = ie.job.Filename
-				daemon.job.FileContent = ie.job.FileContent
-				daemon.job.Status = "Waiting for a button"
-				daemon.job.Progress = ie.job.Progress
-				daemon.job.Owner = ie.job.Owner
+				daemon.job.Id = daemon.ie.job.Id
+				daemon.job.Filename = daemon.ie.job.Filename
+				daemon.job.FileContent = daemon.ie.job.FileContent
+				daemon.job.Progress = daemon.ie.job.Progress
+				daemon.job.Owner = daemon.ie.job.Owner
 
-				if err = ie.reportJobStatusChange(daemon.job); err != nil {
-					log.Error("Can't report it to intern: ", err)
-				}
+				daemon.UpdateStatus("Waiting for a button")
+
 				os.Remove(daemon.buttonfile)
 				os.Remove(daemon.gizmostatusfile)
 
@@ -164,41 +160,32 @@ func main() {
 
 			case "Waiting for a button":
 				log.Info("Job ", daemon.job.Id, " is waiting")
-				err = ie.getJob(daemon.job.Id)
+				err = daemon.ie.getJob(daemon.job.Id)
 				if err != nil {
 					log.Error("Can't get job status from intern: ", err)
 				} else {
-					log.Info("Job status on intern: ", ie.job.Status)
+					log.Info("Job status on intern: ", daemon.ie.job.Status)
 				}
-				if err == nil && ie.job.Status == "Cancelling" {
+				if err == nil && daemon.ie.job.Status == "Cancelling" {
 					log.Info("The job is cancelling")
-					daemon.job.Status = "Cancelling"
+					daemon.UpdateStatus("Cancelling")
 					break
 				}
 				gizmostatusfileStat, err := os.Stat(daemon.gizmostatusfile)
 				if err != nil {
 					log.Info("Job was canceled through device, canceling")
-					daemon.job.Status = "Cancelling"
-					if err = ie.reportJobStatusChange(daemon.job); err != nil {
-						log.Error("Can't report it to intern: ", err)
-					}
+					daemon.UpdateStatus("Cancelling")
 				} else if gizmostatusfileStat.ModTime().Add(waitingForButtonInterval).After(time.Now()) {
 					if daemon.checkButtonPressed() {
-						daemon.job.Status = "Sending to printer"
-						if err = ie.reportJobStatusChange(daemon.job); err != nil {
-							log.Error("Can't report it to intern: ", err)
-						}
+						daemon.UpdateStatus("Sending to printer")
 					} else {
 						log.Info("Waiting ", gizmostatusfileStat.ModTime().Add(waitingForButtonInterval).Unix()-time.Now().Unix(), " more seconds for somebody to press the button")
 					}
 				} else {
 					log.Warning("Nobody pressed the button on time")
-					daemon.job.Status = "Button timeout"
-					if err = ie.reportJobStatusChange(daemon.job); err != nil {
-						log.Error("Can't report it to intern: ", err)
-					}
+					daemon.UpdateStatus("Button timeout")
 					log.Warning("Timeout while waiting for a job. Switching back to ", daemon.job.Status)
-					daemon.job.Status = "Waiting for job"
+					daemon.UpdateStatus("Waiting for job")
 					daemon.job.Id = 0
 					os.Remove(daemon.gizmostatusfile)
 				}
@@ -221,12 +208,7 @@ func main() {
 					log.Error("Failed to create Feeder: ", err)
 					break
 				}
-
-				log.Info("Mark as Printing on intern")
-				daemon.job.Status = "Printing"
-				if err = ie.reportJobStatusChange(daemon.job); err != nil {
-					log.Error("Can't report it to intern: ", err)
-				}
+				daemon.UpdateStatus("Printing")
 
 				go daemon.feeder.Feed()
 
@@ -235,25 +217,19 @@ func main() {
 
 				// Check if status file does not exist (removed through device)
 				if _, err := os.Stat(daemon.gizmostatusfile); os.IsNotExist(err) {
-					daemon.job.Status = "Cancelling"
 					log.Warning("Was canceled through device. Canceling")
 					daemon.feeder.Cancel()
-					if err = ie.reportJobStatusChange(daemon.job); err != nil {
-						log.Error("Can't report progress to intern: ", err)
-					}
+					daemon.UpdateStatus("Cancelling")
 					break
 				}
 
-				err = ie.getJob(daemon.job.Id)
+				err = daemon.ie.getJob(daemon.job.Id)
 				if err != nil {
 					log.Error("Can't report status to intern: ", err)
 				}
-				if err == nil && ie.job.Status == "Cancelling" {
-					daemon.job.Status = "Cancelling"
-					if err = ie.reportJobStatusChange(daemon.job); err != nil {
-						log.Error("Can't report progress to intern: ", err)
-					}
+				if err == nil && daemon.ie.job.Status == "Cancelling" {
 					log.Info("Cancelling the job")
+					daemon.UpdateStatus("Cancelling")
 					daemon.feeder.Cancel()
 					break
 				}
@@ -261,24 +237,22 @@ func main() {
 				daemon.job.feederStatus = daemon.feeder.Status()
 				switch daemon.job.feederStatus {
 				case gcodefeeder.Finished:
-					daemon.job.Status = "Finished"
+					daemon.UpdateStatus("Finished")
 				case gcodefeeder.Error:
-					daemon.job.Status = "Cancelling"
-				}
-				if err = ie.reportJobStatusChange(daemon.job); err != nil {
-					log.Error("Can't report progress to intern: ", err)
+					daemon.UpdateStatus("Cancelling")
 				}
 
 			case "Cancelling":
 				fallthrough
 			case "Finished":
 				log.Info("Deleting from intern")
-				err = ie.deleteJob(daemon.job)
+				err = daemon.ie.deleteJob(daemon.job)
 				if err != nil {
 					log.Error(err)
 				}
 				daemon.job.Id = 0
-				daemon.job.Status = "Waiting for job"
+				daemon.UpdateStatus("Waiting for job")
+
 				// Marking device as free
 				os.Remove(daemon.gizmostatusfile)
 			default:
@@ -286,142 +260,4 @@ func main() {
 			}
 		}
 	}
-}
-
-func (ie *InternEnpoint) reportJobStatusChange(job *Job) error {
-	statusWithProgress := job.Status
-	if job.Status == "Printing" {
-		switch job.feederStatus {
-		case gcodefeeder.Printing:
-			sofar := job.Progress
-			statusWithProgress = fmt.Sprintf("Printing... (%0.1f%%)", sofar)
-		case gcodefeeder.MMUBusy:
-			statusWithProgress = fmt.Sprintf("Printing paused: MMU paused printing")
-		case gcodefeeder.FSensorBusy:
-			statusWithProgress = fmt.Sprintf("Printing paused: Filament sensor paused printing")
-		}
-	}
-
-	data := url.Values{}
-	data.Set("app", ie.Api_app)
-	data.Add("token", ie.Api_key)
-	data.Add("action", "update")
-	data.Add("status", statusWithProgress)
-	data.Add("id", fmt.Sprintf("%d", job.Id))
-	data.Add("printer_name", ie.PrinterName)
-	data.Add("office_name", ie.OfficeName)
-
-	req, err := http.NewRequest("POST", ie.Api_uri+"/job/", bytes.NewBufferString(data.Encode()))
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	return nil
-}
-
-func (ie *InternEnpoint) reschedule() error {
-	data := url.Values{}
-	data.Set("app", ie.Api_app)
-	data.Add("token", ie.Api_key)
-	data.Add("action", "reschedule")
-	data.Add("printer_name", ie.PrinterName)
-	data.Add("office_name", ie.OfficeName)
-
-	req, err := http.NewRequest("POST", ie.Api_uri+"/printer/", bytes.NewBufferString(data.Encode()))
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	return nil
-}
-
-func (ie *InternEnpoint) deleteJob(job *Job) error {
-	data := url.Values{}
-	data.Set("app", ie.Api_app)
-	data.Add("token", ie.Api_key)
-	data.Add("action", "delete")
-	data.Add("id", fmt.Sprintf("%d", job.Id))
-	data.Add("printer_name", ie.PrinterName)
-	data.Add("office_name", ie.OfficeName)
-
-	req, err := http.NewRequest("POST", ie.Api_uri+"/job/", bytes.NewBufferString(data.Encode()))
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	return nil
-}
-
-func (ie *InternEnpoint) nextJob() error {
-	return ie.getJob(0)
-}
-
-func (ie *InternEnpoint) getJob(id int) error {
-	data := url.Values{}
-	data.Set("app", ie.Api_app)
-	data.Add("token", ie.Api_key)
-	data.Add("action", "get")
-	data.Add("printer_name", ie.PrinterName)
-	data.Add("office_name", ie.OfficeName)
-	if id != 0 {
-		data.Add("id", fmt.Sprint(id))
-	}
-
-	req, err := http.NewRequest("POST", ie.Api_uri+"/job/", bytes.NewBufferString(data.Encode()))
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-
-	f, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	var result struct {
-		Success bool
-		Content *Job
-		Error   string
-	}
-
-	err = json.Unmarshal(f, &result)
-	if err != nil {
-		return err
-	}
-	if !result.Success {
-		return fmt.Errorf("job %v action 'get' unsuccessful: %v", id, result.Error)
-	}
-	ie.job = result.Content
-
-	if ie.job.Id == 0 {
-		return errors.New("Nothing to print")
-	}
-
-	return nil
-}
-
-func (ie *InternEnpoint) reportStat() error {
-	data := url.Values{}
-	data.Set("app", ie.Api_app)
-	data.Add("token", ie.Api_key)
-	data.Add("action", "heartbeat")
-	data.Add("printer_name", ie.PrinterName)
-	data.Add("office_name", ie.OfficeName)
-
-	req, err := http.NewRequest("POST", ie.Api_uri+"/printer/", bytes.NewBufferString(data.Encode()))
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	return nil
 }
